+++ src/codemagic/tools/keychain.py
@@ -5,6 +5,7 @@ import json
 import os
 import pathlib
 import shutil
+from datetime import datetime
 from tempfile import NamedTemporaryFile
 from typing import Iterable
 from typing import List
@@ -234,6 +235,24 @@ class Keychain(cli.CliApp, PathFinderMixin):
         if process.returncode != 0:
             raise KeychainError(f'Unable to set {self.path} as default keychain', process)
 
+    @cli.action('use-login')
+    def use_login_keychain(self) -> Keychain:
+        """
+        Use login keychain as the default keychain
+        """
+
+        for keychain_name in ('login.keychain-db', 'login.keychain'):
+            keychain_path = self._keychains_root / keychain_name
+            if keychain_path.is_file():
+                self._path = keychain_path
+                break
+        else:
+            raise KeychainError(f'Login keychain not found from {self._keychains_root}')
+
+        self.logger.info(Colors.GREEN('Use login keychain %s as system default keychain'), self.path)
+        self.make_default()
+        return self
+
     @cli.action('initialize', KeychainArgument.PASSWORD, KeychainArgument.TIMEOUT)
     def initialize(self, password: Password = Password(''), timeout: Optional[Seconds] = None) -> Keychain:
         """
@@ -266,8 +285,15 @@ class Keychain(cli.CliApp, PathFinderMixin):
             self.echo(json.dumps(certificates, sort_keys=True, indent=4))
         return certificates
 
+    @property
+    def _keychains_root(self) -> pathlib.Path:
+        return pathlib.Path('~/Library/Keychains/').expanduser()
+
     def _generate_path(self):
-        with NamedTemporaryFile(prefix='build_', suffix='.keychain') as tf:
+        keychain_dir = self._keychains_root / 'codemagic-cli-tools'
+        keychain_dir.mkdir(parents=True, exist_ok=True)
+        date = datetime.now().strftime('%d-%m-%y')
+        with NamedTemporaryFile(prefix=f'{date}_', suffix='.keychain-db', dir=keychain_dir) as tf:
             self._path = pathlib.Path(tf.name)
 
     @cli.action('add-certificates',